remove BattleAttributes class
merge with ActivePokemon before moving ActivePokemon to PartyPokemon
