+++ rec/generator/interfaces.txt
@@ -1044,7 +1044,7 @@ PowerCountMove:
 		Override: True
 		Header: double getMultiplier(Battle b, ActivePokemon user, ActivePokemon victim)
 		Default:
-			return Math.min(user.getAttributes().getCount(), 5)*(this.doubleDefenseCurled() && user.hasEffect(EffectNamesies.USED_DEFENSE_CURL) ? 2 : 1);
+			return Math.min(user.getCount(), 5)*(this.doubleDefenseCurled() && user.hasEffect(EffectNamesies.USED_DEFENSE_CURL) ? 2 : 1);
 			###
 	*
 ***
