+++ src/battle/attack/Attack.java
@@ -1831,7 +1831,7 @@ public abstract class Attack implements Serializable {
         public void applyDamage(ActivePokemon me, ActivePokemon o, Battle b) {
             o.addEffect((PokemonEffect)EffectNamesies.BRACING.getEffect());
             super.applyDamage(me, o, b);
-            o.getAttributes().removeEffect(EffectNamesies.BRACING);
+            o.removeEffect(EffectNamesies.BRACING);
         }
     }
 
@@ -1913,7 +1913,7 @@ public abstract class Attack implements Serializable {
 
         @Override
         public double getMultiplier(Battle b, ActivePokemon user, ActivePokemon victim) {
-            return user.getAttributes().getCount();
+            return user.getCount();
         }
     }
 
@@ -2021,7 +2021,7 @@ public abstract class Attack implements Serializable {
 
         @Override
         public double getMultiplier(Battle b, ActivePokemon user, ActivePokemon victim) {
-            return user.getAttributes().hasTakenDamage() ? 2 : 1;
+            return user.hasTakenDamage() ? 2 : 1;
         }
     }
 
@@ -2189,7 +2189,7 @@ public abstract class Attack implements Serializable {
 
         @Override
         public void beginAttack(Battle b, ActivePokemon attacking, ActivePokemon defending) {
-            this.mirror = defending.getAttributes().getLastMoveUsed();
+            this.mirror = defending.getLastMoveUsed();
         }
 
         @Override
@@ -3967,7 +3967,7 @@ public abstract class Attack implements Serializable {
 
         @Override
         public boolean applies(Battle b, ActivePokemon user, ActivePokemon victim) {
-            return user.getAttributes().isFirstTurn();
+            return user.isFirstTurn();
         }
     }
 
@@ -4563,7 +4563,7 @@ public abstract class Attack implements Serializable {
 
         @Override
         public double getMultiplier(Battle b, ActivePokemon user, ActivePokemon victim) {
-            return user.getAttributes().hasTakenDamage() ? 2 : 1;
+            return user.hasTakenDamage() ? 2 : 1;
         }
     }
 
@@ -5504,14 +5504,14 @@ public abstract class Attack implements Serializable {
 
         @Override
         public void uniqueEffects(Battle b, ActivePokemon user, ActivePokemon victim) {
-            Move last = victim.getAttributes().getLastMoveUsed();
+            Move last = victim.getLastMoveUsed();
             Messages.add(victim.getName() + "'s " + last.getAttack().getName() + "'s PP was reduced by " + last.reducePP(4) + "!");
         }
 
         @Override
         public boolean applies(Battle b, ActivePokemon user, ActivePokemon victim) {
             // Fails if the victim hasn't attacked yet, their last move already has 0 PP, or they don't actually know the last move they used
-            Move last = victim.getAttributes().getLastMoveUsed();
+            Move last = victim.getLastMoveUsed();
             return last != null && last.getPP() > 0 && victim.hasMove(b, last.getAttack().namesies());
         }
     }
@@ -5817,13 +5817,13 @@ public abstract class Attack implements Serializable {
         @Override
         public void applyDamage(ActivePokemon me, ActivePokemon o, Battle b) {
             // Fails if no damage to reflect or if the opponent isn't using an attack of the proper category
-            int damageTaken = me.getAttributes().getDamageTaken();
+            int damageTaken = me.getDamageTaken();
             o.reduceHealth(b, damageTaken*2);
         }
 
         @Override
         public boolean applies(Battle b, ActivePokemon user, ActivePokemon victim) {
-            return user.getAttributes().getDamageTaken() > 0 && victim.getMove() != null && victim.getAttack().getCategory() == MoveCategory.SPECIAL && !b.isFirstAttack();
+            return user.getDamageTaken() > 0 && victim.getMove() != null && victim.getAttack().getCategory() == MoveCategory.SPECIAL && !b.isFirstAttack();
         }
     }
 
@@ -5842,13 +5842,13 @@ public abstract class Attack implements Serializable {
         @Override
         public void applyDamage(ActivePokemon me, ActivePokemon o, Battle b) {
             // Fails if no damage to reflect or if the opponent isn't using an attack of the proper category
-            int damageTaken = me.getAttributes().getDamageTaken();
+            int damageTaken = me.getDamageTaken();
             o.reduceHealth(b, damageTaken*2);
         }
 
         @Override
         public boolean applies(Battle b, ActivePokemon user, ActivePokemon victim) {
-            return user.getAttributes().getDamageTaken() > 0 && victim.getMove() != null && victim.getAttack().getCategory() == MoveCategory.PHYSICAL && !b.isFirstAttack();
+            return user.getDamageTaken() > 0 && victim.getMove() != null && victim.getAttack().getCategory() == MoveCategory.PHYSICAL && !b.isFirstAttack();
         }
     }
 
@@ -6346,7 +6346,7 @@ public abstract class Attack implements Serializable {
         @Override
         public void startTurn(Battle b, ActivePokemon me) {
             super.applyBasicEffects(b, me, me);
-            me.getAttributes().setReducePP(true);
+            me.setReducePP(true);
         }
 
         @Override
@@ -6419,7 +6419,7 @@ public abstract class Attack implements Serializable {
 
         @Override
         public void endAttack(Battle b, ActivePokemon user, ActivePokemon victim) {
-            user.getAttributes().removeEffect(EffectNamesies.FIDDY_PERCENT_STRONGER);
+            user.removeEffect(EffectNamesies.FIDDY_PERCENT_STRONGER);
         }
 
         @Override
@@ -6748,7 +6748,7 @@ public abstract class Attack implements Serializable {
 
         @Override
         public void beginAttack(Battle b, ActivePokemon attacking, ActivePokemon defending) {
-            this.mirror = defending.getAttributes().getLastMoveUsed();
+            this.mirror = defending.getLastMoveUsed();
         }
 
         @Override
@@ -6868,7 +6868,7 @@ public abstract class Attack implements Serializable {
 
         @Override
         public double getMultiplier(Battle b, ActivePokemon user, ActivePokemon victim) {
-            return user.getAttributes().hasTakenDamage() ? 2 : 1;
+            return user.hasTakenDamage() ? 2 : 1;
         }
     }
 
@@ -7076,7 +7076,7 @@ public abstract class Attack implements Serializable {
         public void beginAttack(Battle b, ActivePokemon attacking, ActivePokemon defending) {
             this.types = new ArrayList<>();
 
-            Move lastMoveUsed = b.getOtherPokemon(attacking).getAttributes().getLastMoveUsed();
+            Move lastMoveUsed = b.getOtherPokemon(attacking).getLastMoveUsed();
             if (lastMoveUsed != null) {
                 Type attackingType = lastMoveUsed.getType();
                 for (Type type : Type.values()) {
@@ -7700,7 +7700,7 @@ public abstract class Attack implements Serializable {
 
         @Override
         public void beginAttack(Battle b, ActivePokemon attacking, ActivePokemon defending) {
-            this.copy = b.getOtherPokemon(attacking).getAttributes().getLastMoveUsed();
+            this.copy = b.getOtherPokemon(attacking).getLastMoveUsed();
         }
 
         @Override
@@ -8051,12 +8051,12 @@ public abstract class Attack implements Serializable {
 
         @Override
         public void applyDamage(ActivePokemon me, ActivePokemon o, Battle b) {
-            o.reduceHealth(b, (int)(me.getAttributes().getDamageTaken()*1.5));
+            o.reduceHealth(b, (int)(me.getDamageTaken()*1.5));
         }
 
         @Override
         public boolean applies(Battle b, ActivePokemon user, ActivePokemon victim) {
-            return user.getAttributes().getDamageTaken() > 0 && !b.isFirstAttack();
+            return user.getDamageTaken() > 0 && !b.isFirstAttack();
         }
     }
 
@@ -9330,7 +9330,7 @@ public abstract class Attack implements Serializable {
                 if (b.isWildBattle()) {
                     victim.removeItem();
                 } else {
-                    user.getAttributes().setCastSource(ItemNamesies.NO_ITEM.getItem());
+                    user.setCastSource(ItemNamesies.NO_ITEM.getItem());
                     EffectNamesies.CHANGE_ITEM.getEffect().apply(b, user, victim, CastSource.CAST_SOURCE, false);
                 }
             }
@@ -10393,7 +10393,7 @@ public abstract class Attack implements Serializable {
 
         @Override
         public boolean applies(Battle b, ActivePokemon user, ActivePokemon victim) {
-            return user.getAttributes().isFirstTurn();
+            return user.isFirstTurn();
         }
     }
 
@@ -10676,7 +10676,7 @@ public abstract class Attack implements Serializable {
 
         @Override
         public boolean applies(Battle b, ActivePokemon user, ActivePokemon victim) {
-            return user.getAttributes().isFirstTurn();
+            return user.isFirstTurn();
         }
     }
 
@@ -11109,7 +11109,7 @@ public abstract class Attack implements Serializable {
 
         @Override
         public void endAttack(Battle b, ActivePokemon user, ActivePokemon victim) {
-            user.getAttributes().removeEffect(EffectNamesies.BREAKS_THE_MOLD);
+            user.removeEffect(EffectNamesies.BREAKS_THE_MOLD);
         }
     }
 
@@ -11129,7 +11129,7 @@ public abstract class Attack implements Serializable {
 
         @Override
         public void endAttack(Battle b, ActivePokemon user, ActivePokemon victim) {
-            user.getAttributes().removeEffect(EffectNamesies.BREAKS_THE_MOLD);
+            user.removeEffect(EffectNamesies.BREAKS_THE_MOLD);
         }
     }
 
@@ -11199,7 +11199,7 @@ public abstract class Attack implements Serializable {
 
         @Override
         public double getMultiplier(Battle b, ActivePokemon user, ActivePokemon victim) {
-            return user.getAttributes().lastMoveSucceeded() ? 2 : 1;
+            return user.lastMoveSucceeded() ? 2 : 1;
         }
     }
