+++ src/battle/effect/generic/CastSource.java
@@ -10,7 +10,7 @@ public enum CastSource {
     ATTACK(false, (b, caster) -> caster.getAttack()),
     ABILITY(true, (b, caster) -> caster.getAbility()),
     HELD_ITEM(true, (b, caster) -> caster.getHeldItem(b)),
-    CAST_SOURCE(false, (b, caster) -> caster.getAttributes().getCastSource()),
+    CAST_SOURCE(false, (b, caster) -> caster.getCastSource()),
     USE_ITEM,
     EFFECT;
