+++ src/battle/effect/generic/EffectInterfaces.java
@@ -1772,7 +1772,7 @@ public final class EffectInterfaces {
 
         @Override
         default double getMultiplier(Battle b, ActivePokemon user, ActivePokemon victim) {
-            return Math.min(user.getAttributes().getCount(), 5)*(this.doubleDefenseCurled() && user.hasEffect(EffectNamesies.USED_DEFENSE_CURL) ? 2 : 1);
+            return Math.min(user.getCount(), 5)*(this.doubleDefenseCurled() && user.hasEffect(EffectNamesies.USED_DEFENSE_CURL) ? 2 : 1);
         }
     }
